[#2] Refactor: mockdata에서 불러온 데이터와 장바구니 데이터가 다른 타입을 갖기 때문에 다로 이름을 붙였습니다
- 타입들은 types폴더로 옮겼습니다
